MyStatusMod/ScreenCapture/requirements.txt
pyobjc
pillow
